requirements.txt
Pillow
streamlit

